Implemented AppTexture, added copy(SDL_Texture*, SDL_Rect*) to AppRenderer
AppTexture represents an image file in an application. The AppTexture currently relies on AppRenderer (through App::instance()->getRenderer()), but this is a candidate for change.

AppRenderer's copy(SDL_Texture*, SDL_Rect*) function would normally be declared using AppTexture instead of SDL_Texture*, but this was not done so as to avoid circular dependencies. It's possible that the other classes in PluSDL will adopt this style (using the lower level SDL structures/classes instead).