Removed unneeded destructors, added width and height logic to AppWindow
Destructors for the classes that had them removed will be fine with the default ones, so there is no need to define them at this moment in time.

To get the height and width of an AppWindow, users had to call SDL_GetWindowSize(). This goes against the projects goal of hiding SDL2 from the end-user, so height and weight was added to the AppWindow class. For non-resizable windows, no updating of the width and height is required since it will remain constant. However, there is an issue when windows have the SDL_WINDOW_RESIZABLE flag passed since the AppWindow's dimension variables are not automatically updated when the window is resized. For now the logic for updating these variables is temporarily located in App's constructor, where an event function is added that waits for SDL_WINDOW_EVENTs and updates the AppWindow accordingly.